create admin layout and clean code
