+++ src/pages/admin/index.tsx
@@ -1,9 +1,6 @@
 import {
   NextPage, GetServerSideProps, GetServerSidePropsContext,
 } from 'next';
-import dynamic from 'next/dynamic';
-
-import BasicLayout from '../../components/BasicLayout';
 
 import { useTranslation } from '~/i18n';
 import { CrowiRequest } from '~/interfaces/crowi-request';
@@ -14,9 +11,10 @@ import ConfigLoader from '~/server/service/config-loader';
 
 import {
   useCurrentUser,
-  useAppTitle, useConfidential,
   useSearchServiceConfigured, useSearchServiceReachable,
 } from '../../stores/context';
+
+import AdminLayout from '~/components/AdminLayout';
 
 
 const pluginUtils = new PluginUtils();
@@ -24,8 +22,6 @@ const pluginUtils = new PluginUtils();
 type Props = CommonProps & {
   currentUser: any,
 
-  appTitle: string,
-  confidential: string,
   isSearchServiceConfigured: boolean,
   isSearchServiceReachable: boolean,
 
@@ -40,42 +36,23 @@ type Props = CommonProps & {
 const AdminHomePage: NextPage<Props> = (props: Props) => {
   const { t } = useTranslation();
   const title = t('Wiki Management Home Page');
-  const AdminNavigation = dynamic(() => import('~/client/js/components/Admin/Common/AdminNavigation'), { ssr: false });
 
   useCurrentUser(props.currentUser != null ? JSON.parse(props.currentUser) : null);
 
-  useAppTitle(props.appTitle);
-  useConfidential(props.confidential);
   useSearchServiceConfigured(props.isSearchServiceConfigured);
   useSearchServiceReachable(props.isSearchServiceReachable);
 
   return (
-    <>
-      <BasicLayout title="GROWI">
-        <header className="py-0">
-          <h1 className="title">{title}</h1>
-        </header>
-        <div id="main" className="main">
-          <div className="container-fluid">
-            <div className="row">
-              <div className="col-lg-3">
-                <AdminNavigation />
-              </div>
-              <div className="col-lg-9">
-                <AdminHome
-                  growiVersion={props.growiVersion}
-                  nodeVersion={props.nodeVersion}
-                  npmVersion={props.npmVersion}
-                  yarnVersion={props.yarnVersion}
-                  installedPlugins={props.installedPlugins}
-                  envVars={props.envVars}
-                />
-              </div>
-            </div>
-          </div>
-        </div>
-      </BasicLayout>
-    </>
+    <AdminLayout title={title}>
+      <AdminHome
+        growiVersion={props.growiVersion}
+        nodeVersion={props.nodeVersion}
+        npmVersion={props.npmVersion}
+        yarnVersion={props.yarnVersion}
+        installedPlugins={props.installedPlugins}
+        envVars={props.envVars}
+      />
+    </AdminLayout>
   );
 };
 
@@ -83,7 +60,7 @@ export const getServerSideProps: GetServerSideProps = async(context: GetServerSi
   const req: CrowiRequest = context.req as CrowiRequest;
   const { crowi } = req;
   const {
-    appService, searchService,
+    searchService,
   } = crowi;
 
   const { user } = req;
@@ -94,7 +71,6 @@ export const getServerSideProps: GetServerSideProps = async(context: GetServerSi
     props.currentUser = JSON.stringify(user.toObject());
   }
 
-  props.confidential = appService.getAppConfidential();
   props.isSearchServiceConfigured = searchService.isConfigured;
   props.isSearchServiceReachable = searchService.isReachable;
